hide progress bar, move adjudicate button
