+++ frontend/src/components/AnnotateCard/index.jsx
@@ -14,7 +14,7 @@ export default function AnnotateCard({ card, labels, onSelectLabel, readonly = f
     return (
         <div className="cardface clearfix">
             <div className="cardface-datacard">
-                <CardData card={card} />
+                <CardData card={card} onSkip={onSkip} />
                 {suggestions && labels.length > 5 && (
                     <SuggestedLabels card={card} labels={labels} onSelectLabel={onSelectLabel} />
                 )}
@@ -33,12 +33,13 @@ function drawEditOptions(card, labels, onSelectLabel, onSkip, onDiscard) {
 
     return (
         <ButtonToolbar className="btn-toolbar">
+            {/* { onSkip != null && drawSkipButton(card, onSkip)} */}
+            <div className="toolbar-gap" />
             {labels.length > 5 ? (
                 drawLabelSelect(card, labelsOptions, onSelectLabel)
             ) : (
                 drawLabelButtons(card, labels, onSelectLabel)
             )}
-            { onSkip != null && drawSkipButton(card, onSkip) }
             { onDiscard != null && drawDiscardButton(card, onDiscard) }
         </ButtonToolbar>
     );
@@ -72,28 +73,6 @@ function drawLabelButtons(card, labels, onSelectLabel) {
                 {opt["name"]}
             </Button>
         ))
-    );
-}
-
-function drawSkipButton(card, onSkip) {
-    return (
-        <OverlayTrigger
-            placement="top"
-            overlay={
-                <Tooltip id="skip_tooltip">
-                    Clicking this button will send this
-                    document to an administrator for review
-                </Tooltip>
-            }
-        >
-            <Button
-                className="ajucate-button"
-                onClick={() => onSkip(card)}
-                variant="info"
-            >
-                Adjudicate
-            </Button>
-        </OverlayTrigger>
     );
 }
