+++ frontend/src/components/DataCard/CardData.jsx
@@ -1,15 +1,45 @@
 import React from "react";
+import {
+    Button,
+    Tooltip,
+    OverlayTrigger
+} from "react-bootstrap";
 
-export default function CardData({ card }) {
+export default function CardData({ card, onSkip }) {
     return (
         <div className="cardface-info">
-            <h2>Card {card.id + 1}</h2>
+            <div className="card-title">
+                <h2>Card {card.id + 1}</h2>
+                {drawSkipButton(card, onSkip)}
+            </div>
             <div className="card-data">
                 <h4>Text to Label</h4>
                 <p>{card.text["text"] || card.text["data"]}</p>
             </div>
             {extractMetadata(card)}
         </div>
+    );
+}
+
+function drawSkipButton(card, onSkip) {
+    return (
+        <OverlayTrigger
+            placement="top"
+            overlay={
+                <Tooltip id="skip_tooltip">
+                    Clicking this button will send this
+                    document to an administrator for review
+                </Tooltip>
+            }
+        >
+            <Button
+                className="ajucate-button"
+                onClick={() => onSkip(card)}
+                variant="info"
+            >
+                Adjudicate
+            </Button>
+        </OverlayTrigger>
     );
 }
